src/tensorrt/tensorrt-6.0.1.5/samples/python/introductory_parser_samples/requirements.txt
numpy
Pillow
pycuda